Created a mercurial repository for the 'hilbert matrix and sine' projects group.
In November 2008 I created 3 projects with the intention to conduct
microbenchmarks that would show the difference in performance between native
C++, C++/CLI running inside a CLR console application and C++/CLI running in a
Windows Forms application when calculating the Hilbert matrix of a given order
or the sine of a given angle using custom implementations. I conducted similar
experiments with matlab and Java, which are not included here.

This repository is intended to track the source code of these 3 projects each in
a different branch.